PDS-688 (#18)
Tracking UI: Release Tracking Interface

For Releases Tracking UI:
Added JQuery datetimepicker plugin.
Created a new Package releases for releases UI component. 

For DOI Tracking UI:
Changed the length of the comment validation from 255 to 1024 basic on
the DB schema. 

For Status Tracking UI:
Changed the length of the comment validation from 255 to 1024 basic on
the DB schema.

When click the cancel button for adding, if no any status for the
product, the page will back to the main page (list of the products).

Fixed the typo. 

The test URLs are:
https://pds-dev2.jpl.nasa.gov/tools/tracking/releases/
https://pds-dev2.jpl.nasa.gov/tools/tracking/statusTracking/
https://pds-dev2.jpl.nasa.gov/tools/tracking/doiTracking/